+++ studio/package.json
@@ -29,8 +29,8 @@
     "@sanity/desk-tool": "^2.1.2",
     "@sanity/vision": "^2.1.0",
     "prop-types": "^15.6",
-    "react": "^16.2",
-    "react-dom": "^16.2",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
     "react-icons": "^3.11.0",
     "react-instagram-embed": "^1.5.0",
     "react-twitter-widgets": "^1.9.5",
@@ -44,8 +44,8 @@
     "eslint-plugin-import": "^2.17.2",
     "eslint-plugin-node": "^9.0.1",
     "eslint-plugin-promise": "^4.1.1",
-    "eslint-plugin-react": "^7.13.0",
-    "eslint-plugin-standard": "^4.0.0",
+    "eslint-plugin-react": "^7.22.0",
+    "eslint-plugin-standard": "^4.1.0",
     "prettier-eslint-cli": "^5.0.0"
   }
 }
