+++ web/src/pages/contact.js
@@ -1,4 +1,4 @@
-import React, { useState, useRef } from "react"
+import React from "react"
 import Layout from "../components/layout"
 import SEO from "../components/seo"
 import { useStaticQuery, graphql } from "gatsby"
