+++ web/src/pages/index.js
@@ -1,153 +1,159 @@
-import React from "react"
-import Layout from "../components/layout"
+import React, { useEffect } from "react"
+import BlogLayout from "../components/blogLayout"
+import { Helmet } from "react-helmet"
 import SEO from "../components/seo"
+import { graphql } from "gatsby"
+import BlogListPage from "../components/BlogListPage"
+import { IoIosCloseCircleOutline } from "react-icons/io"
 
-import { useStaticQuery, graphql } from "gatsby"
-import "../styles/index.css"
-
-const IndexPage = () => {
-  const data = useStaticQuery(graphql`
-    query {
-      sanitySiteSettings {
-        introText
-        introText2
-        siteDesc
-        siteLongTitle
-        siteTitle
-      }
-      sanityHome {
-        offeringsList {
-          content
-          name
-          mainImage {
-            alt
-            asset {
-              url
-            }
-          }
-        }
-        ourWork {
-          name
-          content
-          mainImage {
-            alt
-            asset {
-              url
-            }
-          }
-        }
-      }
-    }
-  `)
+const BlogsPage = ({ data }) => {
+  const [filters, setFilters] = React.useState([])
+  const blogs = data.allBlogs.nodes.sort((a, b) =>
+    a.publishedAt < b.publishedAt ? 1 : -1
+  )
+  const [filteredBlogs, setFilteredBlogs] = React.useState(blogs)
   const siteSettings = data.sanitySiteSettings
-  const offeringsList = data.sanityHome.offeringsList
-  const ourWork = data.sanityHome.ourWork
+  useEffect(() => {
+    setFilteredBlogs(blogs)
+    var newBlogs
+    if (filters !== []) {
+      filters.forEach(filter => {
+        if (filter.type === "tag") {
+          newBlogs = filteredBlogs.filter(blog =>
+            blog.tags.includes(filter.name)
+          )
+          setFilteredBlogs(newBlogs)
+        } else {
+          newBlogs = filteredBlogs.filter(
+            blog => blog.author.name === filter.name
+          )
 
+          setFilteredBlogs(newBlogs)
+        }
+      })
+    }
+  }, [filters])
+
+  function containsObject(newFilter) {
+    var contains = false
+    filters.forEach(item => {
+      if (item.name === newFilter.name) {
+        contains = true
+      }
+    })
+    return contains
+  }
+  const filterData = data => {
+    const newFilter = data
+
+    if (containsObject(newFilter)) {
+    } else {
+      setFilters(filters.concat(newFilter))
+    }
+  }
+
+  const clearFilter = filter => {
+    setFilteredBlogs(blogs)
+    setFilters(filters.filter(filterData => filterData !== filter))
+  }
   return (
-    <Layout bgImg home={true}>
+    <BlogLayout>
+      <Helmet title="Blogs" />
       <SEO
         title={`${siteSettings.siteTitle} | ${siteSettings.siteLongTitle}`}
         description={siteSettings.siteDesc}
       />
-      <section className="sm:mx-5 lg:mx-auto max-w-lg" name="introduction">
-        <h1 className="heroText">{siteSettings.introText}</h1>
-        <h2 className="text-fl-primary font-thin text-3xl">
-          {siteSettings.introText2}
-        </h2>
-      </section>
-      <section name="offerings" className="pt-10 flex justify-center flex-wrap">
-        {offeringsList.map(offering => (
-          <div className="flex max-w-xs items-center flex-col">
-            <div className="flex px-10 pt-6 min-h-full sm:items-center xl:items-start flex-col">
-              <div className="flex h-32 mb-2 items-center">
-                <img
-                  src={offering.mainImage.asset.url}
-                  alt={offering.mainImage.alt}
+      <div className="pb-20">
+        <div
+          className="para-primary m-0 font-semibold"
+          style={{ color: "#E05455" }}
+        >
+          <p className="m-0">Blogs</p>{" "}
+        </div>
+
+        {filters ? (
+          <>
+            {filters.map(filter => (
+              <span
+                className="pl-2 py-1 pr-5 text-xs relative capitalize inline-block bg-white rounded-full md:pr-8 md:pl-5 md:py-2 lg:pr-8 lg:pl-5 lg:py-2 xl:pr-8 xl:pl-5 xl:py-2 md:text-lg lg:text-lg xl:text-lg font-bold text-black mr-2 mb-2"
+                style={{ cursor: "pointer" }}
+                onClick={() => clearFilter(filter)}
+                role="button"
+                tabIndex="0"
+                onKeyPress={() => clearFilter(filter)}
+              >
+                {filter.name}
+                <IoIosCloseCircleOutline
+                  className=" absolute py-1 mb-0 mx-2"
+                  style={{ right: 0 }}
                 />
-              </div>
-              <div className="flex-col sm:text-center xl:text-left">
-                <h2 className="text-md mt-4 text-fl-primary font-medium">
-                  {offering.name}
-                </h2>
-                <div className="leading-normal tracking-normal font-light text-lg w-auto text-white">
-                  {offering.content}
-                </div>
-              </div>
-            </div>
-          </div>
-        ))}
-      </section>
-      <hr
-        className="sm:mx-4 lg:mx-32 my-20"
-        style={{ borderTop: "0.1rem solid white" }}
-        color="white"
-      />
-      {/* <section
-        name="work"
-        className="grid sm:grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-12"
-      >
-        {ourWork.map(work => (
-          <div
-            className="flex justify-center"
-            style={{ backgroundColor: "rgba(255, 255, 255, 0)" }}
-          >
-            <div key={work.name} className="flex p-4">
-              <div className="flex max-w-sm flex-col">
-                <img
-                  alt={work.mainImage.alt}
-                  src={work.mainImage.asset.url}
-                  className="w-full"
-                />
-                <div
-                  className="p-4" // #E05455
-                  style={{
-                    backgroundColor: "white",
-                    color: "#424242",
-                    borderTop: "solid #77878c",
-                    height: "100%",
-                  }}
-                >
-                  <h2 className="text-xl font-semibold py-2">{work.name}</h2>
-                  <p className="p-0 mx-0 max-w-sm text-md leading-normal font-light">
-                    {work.content}
-                  </p>
-                </div>
-              </div>
-            </div>
-          </div>
-        ))}
-      </section> */}
-      <section name="work" className="flex">
-        <div className="flex flex-wrap justify-center items-stretch">
-          {ourWork.map(work => (
-            <div className="flex justify-center p-4" style={{ maxWidth: 300 }}>
-              <div key={work.name} className="flex max-w-sm flex-col">
-                <img
-                  alt={work.mainImage.alt}
-                  src={work.mainImage.asset.url}
-                  className="w-full"
-                />
-                <div // #E05455
-                  className="p-4"
-                  style={{
-                    backgroundColor: "white",
-                    color: "#424242",
-                    borderTop: "solid #77878c",
-                    height: "100%",
-                  }}
-                >
-                  <h2 className="text-xl font-semibold py-2">{work.name}</h2>
-                  <p className="p-0 mx-0 max-w-sm text-md leading-normal font-light">
-                    {work.content}
-                  </p>
-                </div>
-              </div>
-            </div>
+              </span>
+            ))}
+          </>
+        ) : (
+          <></>
+        )}
+        <div className="grid sm:grid-cols-1 md:grid-cols-2 lg:grid-cols-2 xl:grid-cols-3 gap-12">
+          {filteredBlogs.map(blog => (
+            <BlogListPage blog={blog} filterData={filterData} />
           ))}
         </div>
-      </section>
-    </Layout>
+      </div>
+      <SEO />
+    </BlogLayout>
   )
 }
 
-export default IndexPage
+export default BlogsPage
+
+export const query = graphql`
+  {
+    sanitySiteSettings {
+      introText
+      introText2
+      siteDesc
+      siteLongTitle
+      siteTitle
+    }
+    allBlogs: allSanityBlog {
+      nodes {
+        slug {
+          current
+        }
+        title
+        author {
+          name
+          image {
+            asset {
+              url
+            }
+          }
+        }
+        tags
+        body {
+          children {
+            text
+          }
+        }
+        publishedAt(formatString: "MM/DD/YYYY")
+        time: publishedAt(difference: "hours")
+        read_time
+        mainImage {
+          alt
+          asset {
+            fluid {
+              ...GatsbySanityImageFluid
+            }
+            fixed(width: 370, height: 200) {
+              ...GatsbySanityImageFixed
+            }
+            smallImage: fixed(width: 300, height: 170) {
+              ...GatsbySanityImageFixed
+            }
+            url
+          }
+        }
+      }
+    }
+  }
+`
